Refactored homepage into components. Added JSON files to define data to be added to the homepage. Currently the institutions.json is empty to be filled later.
