+++ src/components/home/institutionscarousel/institutions.json
@@ -1,0 +1,12 @@
+[
+    {
+        "name": "",
+        "img_location": "",
+        "href": ""
+    },
+    {
+        "name": "",
+        "img_location": "",
+        "href": ""
+    }
+]
